Move LoadKindEntities to the entities file
